Change to long

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
 
     public static void main(String[] args) {
-        int[] array = {0, 50, 100};
+        long[] array = {0, 50, 100};
         SalesManager salesManager = new SalesManager(array);
         System.out.println(salesManager.max());
     }
